Move analytics consent into store/index.ts, remove separate wrapper file
