+++ crates/desktop/src/lib/store.ts
@@ -1,16 +1,17 @@
 export { disableAgent, enableAgent, getDisabledAgents } from "./store/agents";
 export {
+	getLastSeenWhatsNewVersion,
+	setLastSeenWhatsNewVersion,
+} from "./store/whats-new";
+export {
 	type AnalyticsConsent,
 	getAnalyticsConsent,
 	getConsentAcked,
 	setAnalyticsConsent,
 	setConsentAcked,
-} from "./store/analytics-consent";
-export {
-	getLastSeenWhatsNewVersion,
-	setLastSeenWhatsNewVersion,
-} from "./store/whats-new";
-export { getStore, initStore } from "./store/index";
+	getStore,
+	initStore,
+} from "./store/index";
 export {
 	getIntegrationPreferences,
 	saveIntegrationPreferences,
